pyproject.toml
[tool.poetry]
name = "comments-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
pillow = "^11.0.0"
django-cors-headers = "^4.6.0"
drf-yasg = "^1.21.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
